Update SmallRye Config to 3.13.0
